% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{Title
K-means s}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{# n * p data matrix, n is the data size and p is the dimension}

\item{K}{# K * p positions of centroids, K is the number of clusters and p is the dimension}

\item{M}{# K * p initial centroids}

\item{numIter}{# number of iterations}
}
\value{
Explain return
This function returns a vector of length n containing the indices of cluster that
each points have been assigned to
}
\description{
Title
K-means s
}
\examples{
# Calling the MyKmeans function to sort the rows of a 20 x 50 matrix X
# into 5 clusters, using the first 5 rows of X as the initial centroids.

X <- matrix(rnorm(1000), 20, 50)
Y <- MyKmeans(X, 5, X[1:5, ], 100)
}
